+++ Wrappers/Python/test/test_algorithms.py
@@ -31,6 +31,7 @@ from cil.framework import ImageGeometry
 from cil.framework import AcquisitionGeometry
 from cil.framework import BlockDataContainer
 from cil.framework import BlockGeometry
+from cil.optimisation.utilities import Sampler
 
 from cil.optimisation.utilities import ArmijoStepSizeRule, ConstantStepSize
 from cil.optimisation.operators import IdentityOperator
@@ -62,14 +63,18 @@ from cil.utilities.quality_measures import mae, mse, psnr
 
 # Fast Gradient Projection algorithm for Total Variation(TV)
 from cil.optimisation.functions import TotalVariation
+
+from cil.plugins.ccpi_regularisation.functions import FGP_TV
+import logging
 from testclass import CCPiTestClass
-from utils import has_astra, initialise_tests
+from utils import has_astra
 
 log = logging.getLogger(__name__)
 initialise_tests()
 
 if has_astra:
     from cil.plugins.astra import ProjectionOperator
+
 
 
 if has_cvxpy:
@@ -103,6 +108,7 @@ class TestGD(CCPiTestClass):
         identity = IdentityOperator(ig)
 
         norm2sq = LeastSquares(identity, b)
+
         step_size = norm2sq.L / 3.
 
         alg = GD(initial=initial, objective_function=norm2sq, step_size=step_size,
@@ -110,6 +116,7 @@ class TestGD(CCPiTestClass):
         alg.max_iteration = 1000
         alg.run(1000,verbose=0)
         self.assertNumpyArrayAlmostEqual(alg.x.as_array(), b.as_array())
+
         alg = GD(initial=initial, objective_function=norm2sq, step_size=step_size,
                  atol=1e-9, rtol=1e-6, max_iteration=20, update_objective_interval=2)
         alg.max_iteration = 20
@@ -909,7 +916,6 @@ class TestSIRT(CCPiTestClass):
         sirt = SIRT(initial=tmp_initial, operator=self.Aop,
                     data=self.bop, max_iteration=5)
         sirt.run(5)
-
         x = tmp_initial.copy()
         x_old = tmp_initial.copy()
 
@@ -1027,11 +1033,213 @@ class TestSIRT(CCPiTestClass):
             sirt.x.as_array(), ig.allocate(0.25).as_array(), 3)
 
 
-class TestSPDHG(unittest.TestCase):
-
+class TestSPDHG(CCPiTestClass):
+    def setUp(self):
+        data = dataexample.SIMPLE_PHANTOM_2D.get(size=(20, 20))
+        self.subsets = 10
+        
     @unittest.skipUnless(has_astra, "cil-astra not available")
     def test_SPDHG_vs_PDHG_implicit(self):
         data = dataexample.SIMPLE_PHANTOM_2D.get(size=(128, 128))
+
+        ig = data.geometry
+        ig.voxel_size_x = 0.1
+        ig.voxel_size_y = 0.1
+
+        detectors = ig.shape[0]
+        angles = np.linspace(0, np.pi, 90)
+        ag = AcquisitionGeometry.create_Parallel2D().set_angles(
+            angles, angle_unit='radian').set_panel(detectors, 0.1)
+        # Select device
+        dev = 'cpu'
+
+        Aop = ProjectionOperator(ig, ag, dev)
+
+        sin = Aop.direct(data)
+        partitioned_data = sin.partition(self.subsets, 'sequential')
+        self.A = BlockOperator(
+            *[IdentityOperator(partitioned_data[i].geometry) for i in range(self.subsets)])
+        self.A2 = BlockOperator(
+            *[IdentityOperator(partitioned_data[i].geometry) for i in range(self.subsets)])
+
+        # block function
+        self.F = BlockFunction(*[L2NormSquared(b=partitioned_data[i])
+                                 for i in range(self.subsets)])
+        alpha = 0.025
+        self.G = alpha * FGP_TV()
+
+    def test_SPDHG_defaults_and_setters(self):
+        gamma = 1.
+        rho = .99
+        spdhg = SPDHG(f=self.F, g=self.G, operator=self.A)
+
+        self.assertListEqual(spdhg._norms, [self.A.get_item(i, 0).norm()
+                                           for i in range(self.subsets)])
+        self.assertListEqual(spdhg._prob_weights, [
+                             1/self.subsets] * self.subsets)
+        self.assertEqual(spdhg._sampler._type, 'random_with_replacement')
+        self.assertListEqual(
+            spdhg.sigma, [gamma * rho / ni for ni in spdhg._norms])
+        self.assertEqual(spdhg.tau, min([pi / (si * ni**2) for pi, ni,
+                                         si in zip(spdhg._prob_weights, spdhg._norms, spdhg.sigma)])*(rho / gamma))
+        self.assertNumpyArrayEqual(
+            spdhg.x.as_array(), self.A.domain_geometry().allocate(0).as_array())
+        self.assertEqual(spdhg.max_iteration, 0)
+        self.assertEqual(spdhg.update_objective_interval, 1)
+
+        gamma = 3.7
+        rho = 5.6
+        spdhg.set_step_sizes_from_ratio(gamma, rho)
+        self.assertListEqual(
+            spdhg.sigma, [gamma * rho / ni for ni in spdhg._norms])
+        self.assertEqual(spdhg.tau, min([pi / (si * ni**2) for pi, ni,
+                                         si in zip(spdhg._prob_weights, spdhg._norms, spdhg.sigma)])*(rho / gamma))
+
+        gamma = 1.
+        rho = .99
+        spdhg.set_step_sizes()
+        self.assertListEqual(
+            spdhg.sigma, [gamma * rho / ni for ni in spdhg._norms])
+        self.assertEqual(spdhg.tau, min([pi / (si * ni**2) for pi, ni,
+                                         si in zip(spdhg._prob_weights, spdhg._norms, spdhg.sigma)])*(rho / gamma))
+
+        spdhg.set_step_sizes(sigma=[1]*self.subsets, tau=100)
+        self.assertListEqual(spdhg.sigma, [1]*self.subsets)
+        self.assertEqual(spdhg.tau, 100)
+
+        spdhg.set_step_sizes(sigma=[1]*self.subsets, tau=None)
+        self.assertListEqual(spdhg.sigma, [1]*self.subsets)
+        self.assertEqual(spdhg.tau, min([(pi / (si * ni**2))*(rho / gamma) for pi, ni,
+                                         si in zip(spdhg._prob_weights, spdhg._norms, spdhg.sigma)]))
+
+        spdhg.set_step_sizes(sigma=None, tau=100)
+        self.assertListEqual(spdhg.sigma, [
+                             gamma * rho*pi / (spdhg.tau*ni**2) for ni, pi in zip(spdhg._norms, spdhg._prob_weights)])
+        self.assertEqual(spdhg.tau, 100)
+
+    def test_spdhg_non_default_init(self):
+        spdhg = SPDHG(f=self.F, g=self.G, operator=self.A, sampler=Sampler.random_with_replacement(10, list(np.arange(1, 11)/55.)),
+                      initial=self.A.domain_geometry().allocate(1), max_iteration=1000, update_objective_interval=10)
+
+        self.assertListEqual(spdhg._prob_weights,  list(np.arange(1, 11)/55.))
+        self.assertNumpyArrayEqual(
+            spdhg.x.as_array(), self.A.domain_geometry().allocate(1).as_array())
+        self.assertEqual(spdhg.max_iteration, 1000)
+        self.assertEqual(spdhg.update_objective_interval, 10)
+        
+        with self.assertRaises(ValueError):
+            spdhg = SPDHG(f=self.F, g=self.G, operator=self.A, _norms=[1]*len(self.A), prob_weights=[1/(self.subsets-1)]*(
+                self.subsets-1)+[0], sampler=Sampler.random_with_replacement(10, list(np.arange(1, 11)/55.)))
+            
+        spdhg = SPDHG(f=self.F, g=self.G, operator=self.A, prob_weights=[1/(self.subsets-1)]*(
+                self.subsets-1)+[0],  initial=self.A.domain_geometry().allocate(1), max_iteration=1000, update_objective_interval=10)
+        
+        self.assertListEqual(spdhg._prob_weights,  [1/(self.subsets-1)]*(self.subsets-1)+[0])     
+        self.assertEqual(spdhg._sampler._type, 'random_with_replacement')
+        
+        
+    def test_spdhg_sampler_gives_too_large_index(self):
+        spdhg = SPDHG(f=self.F, g=self.G, operator=self.A, sampler=Sampler.sequential(20),
+                      initial=self.A.domain_geometry().allocate(1), max_iteration=1000, update_objective_interval=10)
+        with self.assertRaises(IndexError):
+            spdhg.run(12)   
+
+    def test_spdhg_deprecated_vargs(self):
+        spdhg = SPDHG(f=self.F, g=self.G, operator=self.A, norms=[
+                      1]*len(self.A), prob=[1/(self.subsets-1)]*(self.subsets-1)+[0])
+
+        self.assertListEqual(self.A.get_norms_as_list(), [1]*len(self.A))
+        self.assertListEqual(spdhg._norms, [1]*len(self.A))
+        self.assertListEqual(spdhg._sampler.prob_weights, [
+                             1/(self.subsets-1)]*(self.subsets-1)+[0])
+        self.assertListEqual(spdhg._prob_weights, [
+                             1/(self.subsets-1)]*(self.subsets-1)+[0])
+
+        with self.assertRaises(ValueError):
+            spdhg = SPDHG(f=self.F, g=self.G, operator=self.A, _norms=[1]*len(self.A), prob=[1/(self.subsets-1)]*(
+                self.subsets-1)+[0], sampler=Sampler.random_with_replacement(10, list(np.arange(1, 11)/55.)))
+            
+            
+        with self.assertRaises(ValueError):
+            spdhg = SPDHG(f=self.F, g=self.G, operator=self.A, _norms=[1]*len(self.A), prob=[1/(self.subsets-1)]*(
+                self.subsets-1)+[0], prob_weights= [1/(self.subsets-1)]*(self.subsets-1)+[0])
+
+        with self.assertRaises(ValueError):
+            spdhg = SPDHG(f=self.F, g=self.G, operator=self.A, sfdsdf=3,  _norms=[
+                          1]*len(self.A), sampler=Sampler.random_with_replacement(10, list(np.arange(1, 11)/55.)))
+
+
+    def test_spdhg_set_norms(self):
+
+        self.A2.set_norms([1]*len(self.A2))
+        spdhg = SPDHG(f=self.F, g=self.G, operator=self.A2)
+        self.assertListEqual(spdhg._norms, [1]*len(self.A2))
+
+    def test_spdhg_check_convergence(self):
+        spdhg = SPDHG(f=self.F, g=self.G, operator=self.A)
+
+        self.assertTrue(spdhg.check_convergence())
+
+        gamma = 3.7
+        rho = 0.9
+        spdhg.set_step_sizes_from_ratio(gamma, rho)
+        self.assertTrue(spdhg.check_convergence())
+
+        gamma = 3.7
+        rho = 100
+        spdhg.set_step_sizes_from_ratio(gamma, rho)
+        self.assertFalse(spdhg.check_convergence())
+
+        spdhg.set_step_sizes(sigma=[1]*self.subsets, tau=100)
+        self.assertFalse(spdhg.check_convergence())
+
+        spdhg.set_step_sizes(sigma=[1]*self.subsets, tau=None)
+        self.assertTrue(spdhg.check_convergence())
+
+        spdhg.set_step_sizes(sigma=None, tau=100)
+        self.assertTrue(spdhg.check_convergence())
+
+    def test_SPDHG_num_subsets_1(self):
+        data = dataexample.SIMPLE_PHANTOM_2D.get(size=(10, 10))
+
+        subsets = 1
+
+        ig = data.geometry
+        ig.voxel_size_x = 0.1
+        ig.voxel_size_y = 0.1
+
+        detectors = ig.shape[0]
+        angles = np.linspace(0, np.pi, 90)
+        ag = AcquisitionGeometry.create_Parallel2D().set_angles(
+            angles, angle_unit='radian').set_panel(detectors, 0.1)
+        # Select device
+        dev = 'cpu'
+
+        Aop = ProjectionOperator(ig, ag, dev)
+
+        sin = Aop.direct(data)
+        partitioned_data = sin.partition(subsets, 'sequential')
+        A = BlockOperator(
+            *[IdentityOperator(partitioned_data[i].geometry) for i in range(subsets)])
+       
+        # block function
+        F = BlockFunction(*[L2NormSquared(b=partitioned_data[i])
+                                 for i in range(subsets)])
+        alpha = 0.025
+        G = alpha * FGP_TV()
+        
+        spdhg = SPDHG(f=F, g=G, operator=A, max_iteration=10, update_objective_interval=10, print_interval=10, log_file=None)
+        
+        spdhg.run(7)   
+        pdhg = PDHG(f=F, g=G, operator=A, max_iteration=10, update_objective_interval=10, print_interval=10, log_file=None)
+        
+        pdhg.run(7)       
+        self.assertNumpyArrayAlmostEqual(pdhg.solution.as_array(), spdhg.solution.as_array(), decimal=3)
+
+    @unittest.skipUnless(has_astra, "cil-astra not available")
+    def test_SPDHG_vs_PDHG_implicit(self):
+
+        data = dataexample.SIMPLE_PHANTOM_2D.get(size=(12, 12))
 
         ig = data.geometry
         ig.voxel_size_x = 0.1
@@ -1082,11 +1290,12 @@ class TestSPDHG(unittest.TestCase):
 
         # Setup and run the PDHG algorithm
         pdhg = PDHG(f=f, g=g, operator=operator, tau=tau, sigma=sigma,
-                    max_iteration=1000,
-                    update_objective_interval=500)
-        pdhg.run(1000,verbose=0)
+                    max_iteration=80,
+                    update_objective_interval=1000)
+        pdhg.run(1000, verbose=0)
 
-        subsets = 10
+        subsets = 5
+
         size_of_subsets = int(len(angles)/subsets)
         # take angles and create uniform subsets in uniform+sequential setting
         list_angles = [angles[i:i+size_of_subsets]
@@ -1110,19 +1319,23 @@ class TestSPDHG(unittest.TestCase):
 
         g = BlockDataContainer(*AD_list)
 
-        # block function
-        F = BlockFunction(*[KullbackLeibler(b=g[i]) for i in range(subsets)])
-        G = alpha * TotalVariation(50, 1e-4, lower=0, warm_start=True)
 
+        ## block function
+        F = BlockFunction(*[KullbackLeibler(b=g[i]) for i in range(subsets)]) 
+        G = alpha * TotalVariation(50, 1e-4, lower=0, warm_start=True) 
         prob = [1/len(A)]*len(A)
+
         spdhg = SPDHG(f=F, g=G, operator=A,
-                      max_iteration=1000,
-                      update_objective_interval=200, prob=prob)
+                      max_iteration=250, sampler=Sampler.random_with_replacement(len(A), seed=2),
+                      update_objective_interval=1000)
+
         spdhg.run(1000, verbose=0)
+
         qm = (mae(spdhg.get_output(), pdhg.get_output()),
               mse(spdhg.get_output(), pdhg.get_output()),
               psnr(spdhg.get_output(), pdhg.get_output())
               )
+
         log.info("Quality measures %r", qm)
 
         np.testing.assert_almost_equal(mae(spdhg.get_output(), pdhg.get_output()),
@@ -1132,7 +1345,8 @@ class TestSPDHG(unittest.TestCase):
 
     @unittest.skipUnless(has_astra, "ccpi-astra not available")
     def test_SPDHG_vs_PDHG_explicit(self):
-        data = dataexample.SIMPLE_PHANTOM_2D.get(size=(128, 128))
+        data = dataexample.SIMPLE_PHANTOM_2D.get(size=(16, 16))
+
 
         ig = data.geometry
         ig.voxel_size_x = 0.1
@@ -1164,7 +1378,7 @@ class TestSPDHG(unittest.TestCase):
             raise ValueError('Unsupported Noise ', noise)
 
         # %% 'explicit' SPDHG, scalar step-sizes
-        subsets = 10
+        subsets = 5
         size_of_subsets = int(len(angles)/subsets)
         # create Gradient operator
         op1 = GradientOperator(ig)
@@ -1197,8 +1411,9 @@ class TestSPDHG(unittest.TestCase):
 
         prob = [1/(2*subsets)]*(len(A)-1) + [1/2]
         spdhg = SPDHG(f=F, g=G, operator=A,
-                      max_iteration=1000,
-                      update_objective_interval=200, prob=prob)
+                      max_iteration=300,
+                      update_objective_interval=300, sampler=Sampler.random_with_replacement(len(A), prob=prob, seed=10))
+
         spdhg.run(1000, verbose=0)
 
         # %% 'explicit' PDHG, scalar step-sizes
@@ -1216,8 +1431,9 @@ class TestSPDHG(unittest.TestCase):
         f = BlockFunction(f1, f2)
         # Setup and run the PDHG algorithm
         pdhg = PDHG(f=f, g=g, operator=operator, tau=tau, sigma=sigma)
-        pdhg.max_iteration = 1000
-        pdhg.update_objective_interval = 200
+        pdhg.max_iteration = 300
+        pdhg.update_objective_interval = 300
+
         pdhg.run(1000, verbose=0)
 
         # %% show diff between PDHG and SPDHG
@@ -1229,11 +1445,13 @@ class TestSPDHG(unittest.TestCase):
               mse(spdhg.get_output(), pdhg.get_output()),
               psnr(spdhg.get_output(), pdhg.get_output())
               )
+
         log.info("Quality measures: %r", qm)
         np.testing.assert_almost_equal(mae(spdhg.get_output(), pdhg.get_output()),
                                        0.00150, decimal=3)
         np.testing.assert_almost_equal(mse(spdhg.get_output(), pdhg.get_output()),
                                        1.68590e-05, decimal=3)
+
 
 
 class TestCallbacks(unittest.TestCase):
@@ -1298,7 +1516,6 @@ class TestCallbacks(unittest.TestCase):
         algo.run(20, callbacks=[EarlyStopping()])
         self.assertEqual(algo.iteration, 15)
 
-
 class TestADMM(unittest.TestCase):
     def setUp(self):
         ig = ImageGeometry(2, 3, 2)
@@ -1347,6 +1564,7 @@ class TestADMM(unittest.TestCase):
         admm_noaxpby = LADMM(f=G, g=F, operator=K, tau=self.tau, sigma=self.sigma,
                              max_iteration=100, update_objective_interval=10)
         admm_noaxpby.run(1, verbose=0)
+
         np.testing.assert_array_almost_equal(
             admm.solution.as_array(), admm_noaxpby.solution.as_array())